% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aeh}
\alias{aeh}
\title{default potential function, \code{L^2/2}}
\usage{
aeh(L, n)
}
\arguments{
\item{L}{interval length.}

\item{n}{total sequence length, required for function signature,
but not used in this example.}
}
\description{
default potential function, \code{L^2/2}
}
