% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg.R
\name{backtrace}
\alias{backtrace}
\title{EXTRACT segements from segmenTier to IRanges}
\usage{
backtrace(ptr)
}
\arguments{
\item{ptr}{pointer of segment ends}
}
\value{
segment blocks
}
\description{
EXTRACT segements from segmenTier to IRanges
}
