% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{backtrace_c}
\alias{backtrace_c}
\title{backtrace breakpoints}
\usage{
backtrace_c(imax)
}
\arguments{
\item{imax}{a \code{NumericVector} containing backtrace vectors.}
}
\description{
backtrace breakpoints
}
