% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg_r.R
\name{backtrace_r}
\alias{backtrace_r}
\title{backtrace function for consseg}
\usage{
backtrace_r(ptr)
}
\arguments{
\item{ptr}{pointer of segment ends.}
}
\value{
breakpoints
}
\description{
backtrace function for consseg
}
