% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg_r.R
\name{bp2seg}
\alias{bp2seg}
\title{convert a vector of breakpoints incl. 1 and n into a list of
adjacent segments, starting at the breakpoints and
ending 1 before the next start/breakpoint}
\usage{
bp2seg(bp, trim.ends = FALSE)
}
\arguments{
\item{bp}{vector of breakpoints.}

\item{trim.ends}{assume that breakpoints are starts of segments,
and the ends are the \code{next breakpoint -1}, EXCEPT for
the last breakpoint, which corresponds to the total sequence length
and is a true end.}
}
\value{
a data.frame with start and end positions of each segment
}
\description{
convert a vector of breakpoints incl. 1 and n into a list of
adjacent segments, starting at the breakpoints and
ending 1 before the next start/breakpoint
}
