% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg.R
\name{calc_dle}
\alias{calc_dle}
\title{Calculates delta(<=i)}
\usage{
calc_dle(i, w, e, segs)
}
\arguments{
\item{i}{current position}

\item{w}{weight function}

\item{e}{potential function}

\item{segs}{starts/ends vector of segments returning segment width}
}
\value{
delta(<=i)
}
\description{
Calculates delta(<=i)
}
