% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg.R
\name{calc_dlov}
\alias{calc_dlov}
\title{Calculates delta^(<i)}
\usage{
calc_dlov(i, w, e, segs)
}
\arguments{
\item{i}{current position}

\item{w}{weight function}

\item{e}{potential function}

\item{segs}{starts/ends vector of segments returning segment width}
}
\value{
delta^(<i)
}
\description{
Calculates delta^(<i)
}
