% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg.R
\name{calc_ds}
\alias{calc_ds}
\title{Calculates delta*(i',i'')}
\usage{
calc_ds(j1, k, w, e, segs)
}
\arguments{
\item{j1}{current span start}

\item{k}{current span end}

\item{w}{weight function}

\item{e}{potential function}

\item{segs}{starts/ends vector of segments returning segment width}
}
\value{
delta*(i',i'')
}
\description{
Calculates delta*(i',i'')
}
