% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg.R
\name{collapse_segments}
\alias{collapse_segments}
\title{Collapse segements S from segmenTier to SM matrix of segmentations,
where breakpoints are indicated by a weight \eqn{w\in 0,{0,\dots,1}}
where weight is 1/total_nr_of_breakpoints or to be determined}
\usage{
collapse_segments(S, w)
}
\arguments{
\item{S}{list of segmentations (breakpoints)}

\item{w}{weightfunction or 1}
}
\description{
Collapse segements S from segmenTier to SM matrix of segmentations,
where breakpoints are indicated by a weight \eqn{w\in 0,{0,\dots,1}}
where weight is 1/total_nr_of_breakpoints or to be determined
}
