% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg_r.R
\name{compileEquation}
\alias{compileEquation}
\title{pre-compile an potential function equation}
\usage{
compileEquation(e = "L*L/2")
}
\arguments{
\item{e}{equation string using the interval length \code{L} and the total
sequence length \code{n} to calculate potentials during the \code{consseg}
recursion, eg. "(L/n)*log(L/n)" to use the negentropy.}
}
\description{
pre-compile an potential function equation
}
