% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg_r.R
\name{consensus}
\alias{consensus}
\title{Calculate consensus segments from a list of segmentation breakpoints}
\usage{
consensus(b, n, w, e, return = "breakpoints", store = FALSE,
  test = FALSE, verb = 1)
}
\arguments{
\item{b}{list of breakpoints of different segmentations}

\item{n}{total sequence length (\code{max(b)} if not provided)}

\item{w}{weights vector, must sum up to 1 or will be normalized}

\item{e}{potential function either a \code{XPtr} pointer to a
pre-compiled function or a string providing \code{Rcpp} code for
this function, eg. \code{"long double my_aeh(int L, int n){return (exp(L/2)-1);}"}. Note, that the function must have this signature, i.e., take two integers as arguments and return a long double.}

\item{return}{return class, simple "breakpoints" or "segments", where
breakpoints are considered the start, and ends are cut one before.}

\item{store}{for debugging: store and return all internal vectors}

\item{test}{for debugging}

\item{verb}{verbosity level, 0 is silent}
}
\description{
Calculate consensus segments from a list of segmentation breakpoints
}
