% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg.R
\name{consensus_old}
\alias{consensus_old}
\title{Calculate consensus segements C from segmenTier segments}
\usage{
consensus_old(RS, w, e)
}
\arguments{
\item{RS, }{raw segmentation object from segmenTier or dataframe}

\item{w}{weight function or 1}

\item{e}{potential function}
}
\value{
The consesus segmentation as IRanges object
}
\description{
Calculate consensus segements C from segmenTier segments
}
