% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg_r.R
\name{consensus_r}
\alias{consensus_r}
\title{Calculate consensus segments from a list of segmentation breakpoints}
\usage{
consensus_r(b, n, w, e = function(L, n) L^2/2, store = FALSE,
  test = FALSE)
}
\arguments{
\item{b}{list of breakpoints of different segmentations}

\item{n}{total sequence length (\code{max(b)} if not provided)}

\item{w}{weights vector, must sum up to 1 or will be normalized}

\item{e}{potential function, taking two arguments: the length \code{L}
of the evaluated interval and the total sequence length \code{n}}

\item{store}{for debugging: store and return all internal vectors}

\item{test}{for debugging: compare the incrementally calculated
Delta with the very slow direct calculation}
}
\description{
Calculate consensus segments from a list of segmentation breakpoints
}
