% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluateEquation}
\alias{evaluateEquation}
\title{evaluate pre-compiled potential functions}
\usage{
evaluateEquation(e, L, n)
}
\arguments{
\item{e}{a \code{XPtr} pointer provided by \code{\link{compileEquation}}.}

\item{L}{numeric, interval length \code{L} when used as a potential
function in the \code{consseg} recursion.}

\item{n}{numeric, total sequence length \code{n} when used as a
potential function in the \code{consseg} recursion.}
}
\description{
evaluate pre-compiled potential functions
}
