% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg.R
\name{extract_ranges}
\alias{extract_ranges}
\title{EXTRACT segements from segmenTier to IRanges}
\usage{
extract_ranges(S)
}
\arguments{
\item{S}{list of segmentations}
}
\value{
An IRanges object of segement blocks
}
\description{
EXTRACT segements from segmenTier to IRanges
}
