% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg_r.R
\name{plot_breaklist}
\alias{plot_breaklist}
\title{simple plot function for a list of segmentation tables}
\usage{
plot_breaklist(blst, n, add = FALSE, length = 0.1, angle = 45,
  code = 3, col = 1, lwd = 2, axis1 = TRUE, axis2 = TRUE, ...)
}
\description{
simple plot function for a list of segmentation tables
}
