% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg.R
\name{plot_Ranges}
\alias{plot_Ranges}
\title{Plot Ranges of segments
Adopted from \href{https://www.bioconductor.org/packages/devel/bioc/vignettes/IRanges/inst/doc/IRangesOverview.pdf}{IRangesOverview}}
\usage{
plot_Ranges(
  segs,
  xlim = segs,
  main = deparse(substitute(segs)),
  col = "lightgrey",
  border = "black",
  sep = 0.5,
  rep = FALSE
)
}
\arguments{
\item{segs}{ranges}

\item{xlim}{x-axis limit}

\item{main}{plot main title}

\item{col}{color for plot}

\item{border}{color for plot borders}

\item{sep}{separator for consecutive segment blocks}

\item{rep}{are repeats of segements to be plotted}
}
\value{
A plot of segments
}
\description{
Plot Ranges of segments
Adopted from \href{https://www.bioconductor.org/packages/devel/bioc/vignettes/IRanges/inst/doc/IRangesOverview.pdf}{IRangesOverview}
}
