% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg.R
\name{plotRanges}
\alias{plotRanges}
\title{Plot Ranges of segments
Adopted from \href{https://www.bioconductor.org/packages/devel/bioc/vignettes/IRanges/inst/doc/IRangesOverview.pdf}{IRangesOverview}}
\usage{
plotRanges(
  x,
  xlim = x,
  main = deparse(substitute(x)),
  col = "black",
  sep = 0.5
)
}
\arguments{
\item{x}{ranges}

\item{xlim}{x-axis limit}

\item{main}{plot main title}

\item{col}{color for plot}

\item{sep}{separator for consecutive segment blocks}
}
\value{
A plot of segments
}
\description{
Plot Ranges of segments
Adopted from \href{https://www.bioconductor.org/packages/devel/bioc/vignettes/IRanges/inst/doc/IRangesOverview.pdf}{IRangesOverview}
}
