% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg_r.R
\name{random_breakpoints}
\alias{random_breakpoints}
\title{random segmentations}
\usage{
random_breakpoints(m = 10, n = 50, lambda = 5)
}
\arguments{
\item{m}{number of segmentations.}

\item{n}{total length of sequence.}

\item{lambda}{mean of poisson distribution to select number of segments.}
}
\description{
random segmentations
}
