% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg.R
\name{scoref}
\alias{scoref}
\title{Calculates Delta(j+1,k)}
\usage{
scoref(j1, k, dl, dle, dlov, drov, dstar)
}
\arguments{
\item{j1}{current span start}

\item{k}{current span end}

\item{dl}{dl list}

\item{dle}{dle list}

\item{dlov}{dlov list}

\item{drov}{drov list}

\item{dstar}{delta star}
}
\value{
Delta(j+1,k)
}
\description{
Calculates Delta(j+1,k)
}
