% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg.R
\name{simulate_ranges}
\alias{simulate_ranges}
\title{Simulate IRanges of segments}
\usage{
simulate_ranges(l, n, s, r, df = FALSE)
}
\arguments{
\item{l}{length of sequence}

\item{n}{number of sequences}

\item{s}{upper bound for number of segmentations per sequence}

\item{r}{repeat first segmentation n times}

\item{df}{return dataframe}
}
\value{
An IRanges object of simulated segments
}
\description{
Simulate IRanges of segments
}
