% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg_r.R
\name{simulate_ranges_r}
\alias{simulate_ranges_r}
\title{Simulate IRanges of segments}
\usage{
simulate_ranges_r(l, n, s, r, df = FALSE)
}
\arguments{
\item{l}{length of sequence}

\item{n}{average number of segments per segmentation}

\item{s}{upper bound for number of segmentations per sequence}

\item{r}{repeat first segmentation n times}

\item{df}{return dataframe}
}
\value{
An IRanges object of simulated segments
}
\description{
Simulate IRanges of segments
}
