% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sset.R
\docType{data}
\name{sset}
\alias{sset}
\title{Segements from transcriptome time-series from budding yeast.}
\description{
segmenTier processed transcriptome time-series data from a region encompassing
four genes and a regulatory upstream non-coding RNA in budding yeast.
The data set is described in more detail in the publication
Machne, Murray & Stadler (2017) \doi{10.1038/s41598-017-12401-8}.
}
