% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg_r.R
\name{bp2seg}
\alias{bp2seg}
\title{convert a vector of breakpoints incl. 1 and n into a list of
adjacent segments, starting at the breakpoints and
ending 1 before the next start/breakpoint}
\usage{
bp2seg(bp)
}
\arguments{
\item{bp}{vector of breakpoints}
}
\value{
a data.frame with start and end positions of each segment
}
\description{
convert a vector of breakpoints incl. 1 and n into a list of
adjacent segments, starting at the breakpoints and
ending 1 before the next start/breakpoint
}
