% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg_r.R
\name{consensus}
\alias{consensus}
\title{Calculate consensus segments from a list of segmentation breakpoints}
\usage{
consensus(b, n, return = "breakpoints", store = FALSE, test = FALSE)
}
\arguments{
\item{b}{list of breakpoints of different segmentations}

\item{n}{total sequence length (\code{max(b)} if not provided)}

\item{return}{return class of}

\item{store}{for debugging: store and return all internal vectors}

\item{test}{for debugging}
}
\description{
Calculate consensus segments from a list of segmentation breakpoints
}
