% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg.R
\name{consensus}
\alias{consensus}
\title{Calculate consensus segements C from segmenTier segments}
\usage{
consensus(RS)
}
\arguments{
\item{RS, }{raw segmentation object from segmenTier}
}
\description{
Calculate consensus segements C from segmenTier segments
}
