% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg_r.R
\name{consensus_r}
\alias{consensus_r}
\title{Calculate consensus segments from a list of segmentation breakpoints}
\usage{
consensus_r(b, n, w = function(m) return(1/m), aeh = function(L) L^5/5,
  store.matrix = FALSE, test.slow = FALSE)
}
\arguments{
\item{b}{list of breakpoints of different segmentations}

\item{n}{total sequence length (\code{max(b)} if not provided)}

\item{w}{weight function, taking one argument: the index \code{m} of
the respective segmentation in the breakpoint list \code{b}}

\item{store.matrix}{for debugging: store and return all internal vectors}

\item{test.slow}{for debugging: compare the incrementally calculated
Delta with the very slow direct calculation}

\item{e}{potential function, taking one argument: the length \code{L}
of the evaluated interval}
}
\description{
Calculate consensus segments from a list of segmentation breakpoints
}
