% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg.R
\name{extract_segments}
\alias{extract_segments}
\title{EXTRACT segements from segmenTier to ordered data.frame SO}
\usage{
extract_segments(S)
}
\arguments{
\item{S}{list of segmentations (breakpoints)}
}
\description{
EXTRACT segements from segmenTier to ordered data.frame SO
}
