% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consseg_r.R
\name{plot_breaklist}
\alias{plot_breaklist}
\title{Simple plot function for a list of segmentation
breakpoints or segment tables.}
\usage{
plot_breaklist(blst, n, add = FALSE, length = 0.05, angle = 45,
  code = 3, col = 1, lwd = 2, xlab = "position", axis1 = TRUE,
  axis2 = TRUE, ...)
}
\arguments{
\item{blst}{a vector of breakpoints (output from \code{\link{consensus}}),
a named list of breakpoint vectors, a named list of \code{data.frame}s
with start and end columns, or a segmenTier results object
(\code{class(blst)=="segments"}).}

\item{n}{total sequence length (\code{max(unlist(blst))} if not provided).}

\item{add}{add to existing plot.}

\item{length}{length of the edges of the arrow head (in inches).}

\item{angle}{angle from the shaft of the arrow to the edge of the arrow
head.}

\item{code}{integer code, determining \emph{kind} of arrows to be drawn.}

\item{col}{arrow colors, a single color or a vector for each segmentation.}

\item{lwd}{arrow line width.}

\item{xlab}{x-axis label.}

\item{axis1}{draw x-axis.}

\item{axis2}{draw y-axis.}

\item{...}{further arguments to \code{\link{arrows}}.}
}
\description{
Simple plot function for a list of segmentation
breakpoints or segment tables.
}
